Passive Sensor Part 1
The passive sensor component, as well as some of the ship data relating
to them has been finished. Armor and Engine costs were converted to
decimal to keep consistent with pre-existing ship costs, and an
additional test was put into ship tests.

The next part will be putting the rest of the ship,taskforce, and
faction level work for sensors in.

ActiveSensor is just a stub for right now.

+++ Pulsar4X/Pulsar4X.Lib/Entities/Components/Armor.cs
@@ -61,8 +61,8 @@ namespace Pulsar4X.Entities.Components
         /// Cost of the Armor. Ship costs, repair cost, and material resource costs will depend on this. It is naively equal to area in terms of per numbers, but resources required
         /// will vary with tech level. In Aurora costs shift from duranium to neutronium for example.
         /// </summary>
-        private double Cost;
-        public double cost
+        private decimal Cost;
+        public decimal cost
         {
             get { return Cost; }
         }
@@ -83,7 +83,7 @@ namespace Pulsar4X.Entities.Components
 	    public ArmorDefTN()
 	    {
 		    Size = 0.0;
-		    Cost = 0.0;
+		    Cost = 0.0m;
 		    Area = 0.0;
 	    }
 
@@ -162,7 +162,7 @@ namespace Pulsar4X.Entities.Components
 		    }
 
 		    Area = ( area * Depth ) / 4.0;
-		    Cost = Area;
+		    Cost = (decimal)Area;
 		    CNum = (ushort)Math.Floor( strengthReq / (double)Depth );
 	    }
         /// <summary>
